Use DOUBLE for Amount in physical tests

+++ physicalTests/DynamicKsqlGenerationTests.cs
@@ -293,7 +293,7 @@ public class DynamicKsqlGenerationTests
         public int CustomerId { get; set; }
         public int Id { get; set; }
         public string Region { get; set; } = string.Empty;
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public bool IsHighPriority { get; set; }
         public int Count { get; set; }
     }
@@ -314,13 +314,13 @@ public class DynamicKsqlGenerationTests
     {
         public int? CustomerId { get; set; }
         public string Region { get; set; } = string.Empty;
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
     }
 
     public class NullableKeyOrder
     {
         public int? CustomerId { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
     }
 
     public class DummyContext : KsqlContext
@@ -342,15 +342,15 @@ public class DynamicKsqlGenerationTests
             CustomerId = 1,
             Id = 1,
             Region = "east",
-            Amount = 10m,
+            Amount = 10d,
             IsHighPriority = false,
             Count = 1
         });
 
         await ctx.Set<Customer>().AddAsync(new Customer { Id = 1, Name = "alice" });
         await ctx.Set<EventLog>().AddAsync(new EventLog { Level = 1, Message = "init" });
-        await ctx.Set<NullableOrder>().AddAsync(new NullableOrder { CustomerId = 1, Region = "east", Amount = 10m });
-        await ctx.Set<NullableKeyOrder>().AddAsync(new NullableKeyOrder { CustomerId = 1, Amount = 10m });
+        await ctx.Set<NullableOrder>().AddAsync(new NullableOrder { CustomerId = 1, Region = "east", Amount = 10d });
+        await ctx.Set<NullableKeyOrder>().AddAsync(new NullableKeyOrder { CustomerId = 1, Amount = 10d });
 
         await Task.Delay(500);
         await ctx.DisposeAsync();
